Run parts of the publish workflow on PRs
This runs `npm run build` for pull requests so that any issues can be
mitigated before they're merged to the `main` branch. For simplicity of
maintenance the existing "Publish website" workflow is reused instead of
creating a new workflow so that there is a single workflow definition
that needs to be maintained. The publishing of the website is skipped
for pull requests using the conditional on the event type being "push",
so that the publishing to GitHub pages is skipped for pull requests.

The existing permissions block is kept as is and should apply only to
pushes to the `main` branch. This is safe as long as "Send write tokens
to workflows from pull requests" is not configured in the repository
settings. See: https://docs.github.com/en/actions/using-jobs/assigning-permissions-to-jobs#overview