Checks workflow for pull requests
This runs `npm run build` for pull requests so that any issues can be
mitigated before they're merged to the `main` branch. The existing
workflow was copied and adapted to trigger on pull requests. Having the
GitHub pages environment defined in the publish workflow it was not
possible to reuse the existing workflow due to security implications of
trying to allow publishing to GitHub pages from pull requests.