Edits test overview page.
This PR largely rewrites the overview page for testing applications.
The old draft was outdated and pretty hard to follow. This one is updated and, I believe, reader-focused.
This also deletes a few periods from the EC doc that could've led users to accidentally adding typos into their integration test for the EC.