Clean up the top of the secrets how-to-guide
This is now a general doc about uploading secrets.

The main purpose of the change here is to kind of push the information
about the name of the snyk secret into an aside, a note, so that it
doesn't get in the main way of the thrust of the document.

It's still important. Users still need to know it. But, it's not central
to the doc.

Also, I tried to improve the explanatory language around the snyk secret.